Implemented deleteall tweets flow
